Adding a TCK process reminder to the release checklist
This is to prevent a repeat of issues in previous releases, where
people would work on a set of tests without notifying the community
until they were finished.

This resulted in effort duplication and a slower release.

Also, Windows doesn't allow 2 people to access the same machine
concurrently, so this locks up a machine with no way for a user
to (independantly) tell why it's already in use, and whether it's
safe to remove their predecessor.

Signed-off-by: Adam Farley <adfarley@redhat.com>